lightnvm: pblk: implement a write pointer for mapping sectors
In stead of calculating each ppa address based on the sector offset
within a line on the write path, store a write pointer instead
and increment this through the writable sectors in the line.

By using a write pointer we avoid address conversions (and the
divisions involved) from paddrs to ppas.

This prepares pblk for the removal of the mapping bitmap.

Signed-off-by: Hans Holmberg <hans.holmberg@cnexlabs.com>