lightnvm: pblk: remove mapping bitmap
Remove the mapping bitmap, now that we can map sectors without
it.

This includes removing pblk_dealloc_page. The only use of
pblk_dealloc_page is when writing end metadata and seems to
be intended for re-trying i/o submission which does not
make sense anymore. If we have a write failure during emeta
writing, we cannot re-issue writes(as we don't have any
"spare" sectors for emeta to write to), but that's fine as
we can recover the L2P mapping from the metadata until the
whole line has been emergency garbage collected.

Also mop up the no-longer-needed bitmap calculation helpers,
and the bitmap pool.

Signed-off-by: Hans Holmberg <hans.holmberg@cnexlabs.com>